refactor: move Breadcrumbs into SharedComponents
