+++ src/components/HeaderLesson/HeaderWrapper.tsx
@@ -1,8 +1,10 @@
 import { FC } from "react";
 
-import Breadcrumbs, { Breadcrumb } from "../Breadcrumbs";
 import { Hr } from "../Typography";
 
+import Breadcrumbs, {
+  Breadcrumb,
+} from "@/components/SharedComponents/Breadcrumbs/Breadcrumbs";
 import Box from "@/components/SharedComponents/Box";
 import Flex from "@/components/SharedComponents/Flex";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
