+++ src/components/Lesson/LessonDownloads/LessonDownloads.page.tsx
@@ -13,7 +13,7 @@ import {
   ResourceFormProps,
   DownloadResourceType,
 } from "@/components/DownloadAndShareComponents/downloadAndShare.types";
-import Breadcrumbs from "@/components/Breadcrumbs";
+import Breadcrumbs from "@/components/SharedComponents/Breadcrumbs";
 import DownloadCardGroup from "@/components/DownloadAndShareComponents/DownloadCardGroup/DownloadCardGroup";
 import debouncedSubmit from "@/components/DownloadAndShareComponents/helpers/downloadDebounceSubmit";
 import useAnalyticsPageProps from "@/hooks/useAnalyticsPageProps";
