+++ src/components/Posts/PostListing.tsx
@@ -13,7 +13,7 @@ import PostListAndCategories from "@/components/Posts/PostListAndCategories";
 import {
   CrumbPageVariant,
   getBlogWebinarListBreadcrumbs,
-} from "@/components/Breadcrumbs/getBreadcrumbs";
+} from "@/components/SharedComponents/Breadcrumbs/getBreadcrumbs";
 import SummaryCard from "@/components/Card/SummaryCard";
 import Layout from "@/components/Layout";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
@@ -28,7 +28,7 @@ import {
   WebinarListingPageProps,
   webinarToPostListItem,
 } from "@/components/pages/WebinarsIndex.page";
-import Breadcrumbs from "@/components/Breadcrumbs/Breadcrumbs";
+import Breadcrumbs from "@/components/SharedComponents/Breadcrumbs/Breadcrumbs";
 
 type PostListingProps = {
   seo: SeoProps;
