+++ src/components/Posts/PostSingleLayout.tsx
@@ -12,7 +12,9 @@ import Grid, { GridArea } from "@/components/Grid";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import MobileFilters from "@/components/MobileFilters";
 import { Heading } from "@/components/Typography";
-import Breadcrumbs, { Breadcrumb } from "@/components/Breadcrumbs/Breadcrumbs";
+import Breadcrumbs, {
+  Breadcrumb,
+} from "@/components/SharedComponents/Breadcrumbs/Breadcrumbs";
 import Svg from "@/components/Svg/Svg";
 
 type PostSingleLayoutProps = {
