+++ src/components/SharedComponents/Breadcrumbs/Breadcrumbs.tsx
@@ -1,11 +1,11 @@
 import { FC } from "react";
 import styled from "styled-components";
 
-import { BreadcrumbJsonLd } from "../../browser-lib/seo/getJsonLd";
-import Icon from "../Icon";
-import UL from "../Typography/UL";
-import OakLink from "../OakLink";
-import { MaybeOakHref, ResolveOakHrefProps } from "../../common-lib/urls";
+import { BreadcrumbJsonLd } from "../../../browser-lib/seo/getJsonLd";
+import Icon from "../../Icon";
+import UL from "../../Typography/UL";
+import OakLink from "../../OakLink";
+import { MaybeOakHref, ResolveOakHrefProps } from "../../../common-lib/urls";
 
 import ellipsis from "@/styles/ellipsis";
