+++ src/components/pages/CurriculumInfo/CurriculumHeader/CurriculumHeader.tsx
@@ -5,7 +5,7 @@ import Box from "@/components/SharedComponents/Box";
 import Flex from "@/components/SharedComponents/Flex";
 import { Heading, Hr, P } from "@/components/Typography";
 import SubjectIcon from "@/components/SubjectIcon/SubjectIcon";
-import Breadcrumbs from "@/components/Breadcrumbs/Breadcrumbs";
+import Breadcrumbs from "@/components/SharedComponents/Breadcrumbs/Breadcrumbs";
 import TabularNav from "@/components/TabularNav/TabularNav";
 import SubjectPhasePicker, {
   SubjectPhasePickerData,
